move classpath classes to dedicated pkg
